added 101-print_listint_safe.c

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -10,7 +10,7 @@
 
 size_t print_listint_safe(const listint_t *head)
 {
-	size_t i = 0;
+	int i = 0;
 	const listint_t *current, *temp;
 
 	if (head == NULL)
